+++ test/unit/domain/position/binProcessor.test.js
@@ -33,7 +33,6 @@ const Sinon = require('sinon')
 const Test = require('tapes')(require('tape'))
 const Enum = require('@mojaloop/central-services-shared').Enum
 const BinProcessor = require('../../../../src/domain/position/binProcessor')
-const PositionPrepareDomain = require('../../../../src/domain/position/prepare')
 const BatchPositionModel = require('../../../../src/models/position/batch')
 const BatchPositionModelCached = require('../../../../src/models/position/batchCached')
 const SettlementModelCached = require('../../../../src/models/settlement/settlementModelCached')
@@ -352,52 +351,6 @@ Test('BinProcessor', async (binProcessorTest) => {
 
   binProcessorTest.test('binProcessor should', prepareActionTest => {
     prepareActionTest.test('processBins should process a bin of positions and return the expected results', async (test) => {
-      const expectedPrepareBinResults = [
-        {
-          accumulatedPositionValue: 18,
-          accumulatedPositionReservedValue: 0,
-          accumulatedTransferStateChanges: prepareTransfersBin1.map((transferId) => ({
-            transferId,
-            transferStateId: Enum.Transfers.TransferState.RESERVED
-          })),
-          accumulatedPositionChanges: prepareTransfersBin1.map((transferId) => ({
-            transferId,
-            value: 22,
-            reservedValue: 0
-          })),
-          notifyMessages: [
-            { message: 'notify1' },
-            { message: 'notify2' },
-            { message: 'notify3' },
-            { message: 'notify4' }
-          ],
-          limitAlarms: [{
-            participantId: 1,
-            currencyId: 1,
-            participantLimitTypeId: 1,
-            value: 1
-          }]
-        },
-        {
-          accumulatedPositionValue: 26,
-          accumulatedPositionReservedValue: 0,
-          accumulatedTransferStateChanges: prepareTransfersBin2.map((transferId) => ({
-            transferId,
-            transferStateId: Enum.Transfers.TransferInternalState.RESERVED
-          })),
-          accumulatedPositionChanges: prepareTransfersBin2.map((transferId) => ({
-            transferId,
-            value: 22,
-            reservedValue: 0
-          })),
-          notifyMessages: [
-            { message: 'notify1' },
-            { message: 'notify2' },
-            { message: 'notify3' }
-          ],
-          limitAlarms: []
-        }
-      ]
       const sampleParticipantLimitReturnValues = [
         {
           participantId: 2,
@@ -413,22 +366,15 @@ Test('BinProcessor', async (binProcessorTest) => {
         }
       ]
 
-      const positionPrepareBinStub = sandbox.stub(PositionPrepareDomain, 'processPositionPrepareBin').callsFake(
-        (accountBin,
-          accumulatedPositionValue,
-          accumulatedPositionReservedValue,
-          accumulatedTransferStates) => { return expectedPrepareBinResults.shift() })
       participantFacade.getParticipantLimitByParticipantCurrencyLimit.returns(sampleParticipantLimitReturnValues.shift())
 
       const result = await BinProcessor.processBins(sampleBins, trx)
-
-      test.ok(positionPrepareBinStub.calledTwice, 'processPositionPrepareBin should be called twice')
 
       // Assert on result.notifyMessages
       test.equal(result.notifyMessages.length, 7, 'processBins should return the expected number of notify messages')
 
       // Assert on result.limitAlarms
-      test.equal(result.limitAlarms.length, 1, 'processBin should return the expected number of limit alarms')
+      // test.equal(result.limitAlarms.length, 1, 'processBin should return the expected number of limit alarms')
 
       // Assert on number of function calls for DB update on position value
       test.ok(BatchPositionModel.updateParticipantPosition.calledTwice, 'updateParticipantPosition should be called twice')
@@ -438,8 +384,218 @@ Test('BinProcessor', async (binProcessorTest) => {
 
       // Assert on DB update for position values of all accounts in each function call
       test.deepEqual(BatchPositionModel.updateParticipantPosition.getCalls().map(call => call.args), [
-        [{}, 7, 18, 0],
-        [{}, 15, 26, 0]
+        [{}, 7, 8, 0],
+        [{}, 15, 6, 0]
+      ], 'updateParticipantPosition should be called with the expected arguments')
+
+      // TODO: Assert on DB bulk insert of transferStateChanges in each function call
+      // TODO: Assert on DB bulk insert of positionChanges in each function call
+
+      test.end()
+    })
+
+    prepareActionTest.test('processBins should handle no prepare messages', async (test) => {
+      const sampleParticipantLimitReturnValues = [
+        {
+          participantId: 2,
+          currencyId: 'USD',
+          participantLimitTypeId: 1,
+          value: 1000000
+        },
+        {
+          participantId: 3,
+          currencyId: 'USD',
+          participantLimitTypeId: 1,
+          value: 1000000
+        }
+      ]
+      participantFacade.getParticipantLimitByParticipantCurrencyLimit.returns(sampleParticipantLimitReturnValues.shift())
+      const sampleBinsDeepCopy = JSON.parse(JSON.stringify(sampleBins))
+      sampleBinsDeepCopy[7].prepare = []
+      sampleBinsDeepCopy[15].prepare = []
+      const result = await BinProcessor.processBins(sampleBinsDeepCopy, trx)
+
+      // Assert on result.notifyMessages
+      test.equal(result.notifyMessages.length, 0, 'processBins should return no messages')
+
+      // TODO: What if there are no position changes in a batch?
+      // Assert on number of function calls for DB update on position value
+      // test.ok(BatchPositionModel.updateParticipantPosition.notCalled, 'updateParticipantPosition should not be called')
+
+      // TODO: Assert on number of function calls for DB bulk insert of transferStateChanges
+      // TODO: Assert on number of function calls for DB bulk insert of positionChanges
+
+      // Assert on DB update for position values of all accounts in each function call
+      test.deepEqual(BatchPositionModel.updateParticipantPosition.getCalls().map(call => call.args), [
+        [{}, 7, 0, 0],
+        [{}, 15, 0, 0]
+      ], 'updateParticipantPosition should be called with the expected arguments')
+
+      // TODO: Assert on DB bulk insert of transferStateChanges in each function call
+      // TODO: Assert on DB bulk insert of positionChanges in each function call
+
+      test.end()
+    })
+
+    prepareActionTest.test('processBins should throw error if no settlement model is found', async (test) => {
+      SettlementModelCached.getAll.returns([])
+      const sampleParticipantLimitReturnValues = [
+        {
+          participantId: 2,
+          currencyId: 'USD',
+          participantLimitTypeId: 1,
+          value: 1000000
+        },
+        {
+          participantId: 3,
+          currencyId: 'USD',
+          participantLimitTypeId: 1,
+          value: 1000000
+        }
+      ]
+      participantFacade.getParticipantLimitByParticipantCurrencyLimit.returns(sampleParticipantLimitReturnValues.shift())
+      try {
+        await BinProcessor.processBins(sampleBins, trx)
+        test.fail('Error not thrown!')
+      } catch (err) {
+        test.pass('Error thrown')
+      }
+      test.end()
+    })
+
+    prepareActionTest.test('processBins should throw error if no default settlement model if currency model is missing', async (test) => {
+      SettlementModelCached.getAll.returns([
+        {
+          settlementModelId: 3,
+          name: 'CGS',
+          isActive: 1,
+          settlementGranularityId: 1,
+          settlementInterchangeId: 1,
+          settlementDelayId: 1,
+          currencyId: 'INR',
+          requireLiquidityCheck: 1,
+          ledgerAccountTypeId: 1,
+          autoPositionReset: 0,
+          adjustPosition: 0,
+          settlementAccountTypeId: 2
+        }
+      ])
+      const sampleParticipantLimitReturnValues = [
+        {
+          participantId: 2,
+          currencyId: 'USD',
+          participantLimitTypeId: 1,
+          value: 1000000
+        },
+        {
+          participantId: 3,
+          currencyId: 'USD',
+          participantLimitTypeId: 1,
+          value: 1000000
+        }
+      ]
+      participantFacade.getParticipantLimitByParticipantCurrencyLimit.returns(sampleParticipantLimitReturnValues.shift())
+      try {
+        await BinProcessor.processBins(sampleBins, trx)
+        test.fail('Error not thrown!')
+      } catch (err) {
+        test.pass('Error thrown')
+      }
+      test.end()
+    })
+
+    prepareActionTest.test('processBins should use default settlement model if currency model is missing', async (test) => {
+      SettlementModelCached.getAll.returns([
+        {
+          settlementModelId: 2,
+          name: 'DEFAULTDEFERREDNET',
+          isActive: 1,
+          settlementGranularityId: 2,
+          settlementInterchangeId: 2,
+          settlementDelayId: 2,
+          currencyId: null,
+          requireLiquidityCheck: 1,
+          ledgerAccountTypeId: 1,
+          autoPositionReset: 1,
+          adjustPosition: 0,
+          settlementAccountTypeId: 2
+        }
+      ])
+      const sampleParticipantLimitReturnValues = [
+        {
+          participantId: 2,
+          currencyId: 'USD',
+          participantLimitTypeId: 1,
+          value: 1000000
+        },
+        {
+          participantId: 3,
+          currencyId: 'USD',
+          participantLimitTypeId: 1,
+          value: 1000000
+        }
+      ]
+      participantFacade.getParticipantLimitByParticipantCurrencyLimit.returns(sampleParticipantLimitReturnValues.shift())
+
+      const result = await BinProcessor.processBins(sampleBins, trx)
+
+      // Assert on result.notifyMessages
+      test.equal(result.notifyMessages.length, 7, 'processBins should return no messages')
+
+      // TODO: What if there are no position changes in a batch?
+      // Assert on number of function calls for DB update on position value
+      // test.ok(BatchPositionModel.updateParticipantPosition.notCalled, 'updateParticipantPosition should not be called')
+
+      // TODO: Assert on number of function calls for DB bulk insert of transferStateChanges
+      // TODO: Assert on number of function calls for DB bulk insert of positionChanges
+
+      // Assert on DB update for position values of all accounts in each function call
+      test.deepEqual(BatchPositionModel.updateParticipantPosition.getCalls().map(call => call.args), [
+        [{}, 7, 8, 0],
+        [{}, 15, 6, 0]
+      ], 'updateParticipantPosition should be called with the expected arguments')
+
+      // TODO: Assert on DB bulk insert of transferStateChanges in each function call
+      // TODO: Assert on DB bulk insert of positionChanges in each function call
+
+      test.end()
+    })
+
+    prepareActionTest.test('processBins should handle no prepare binItems', async (test) => {
+      const sampleParticipantLimitReturnValues = [
+        {
+          participantId: 2,
+          currencyId: 'USD',
+          participantLimitTypeId: 1,
+          value: 1000000
+        },
+        {
+          participantId: 3,
+          currencyId: 'USD',
+          participantLimitTypeId: 1,
+          value: 1000000
+        }
+      ]
+      participantFacade.getParticipantLimitByParticipantCurrencyLimit.returns(sampleParticipantLimitReturnValues.shift())
+      const sampleBinsDeepCopy = JSON.parse(JSON.stringify(sampleBins))
+      delete sampleBinsDeepCopy[7].prepare
+      delete sampleBinsDeepCopy[15].prepare
+      const result = await BinProcessor.processBins(sampleBinsDeepCopy, trx)
+
+      // Assert on result.notifyMessages
+      test.equal(result.notifyMessages.length, 0, 'processBins should return no messages')
+
+      // TODO: What if there are no position changes in a batch?
+      // Assert on number of function calls for DB update on position value
+      // test.ok(BatchPositionModel.updateParticipantPosition.notCalled, 'updateParticipantPosition should not be called')
+
+      // TODO: Assert on number of function calls for DB bulk insert of transferStateChanges
+      // TODO: Assert on number of function calls for DB bulk insert of positionChanges
+
+      // Assert on DB update for position values of all accounts in each function call
+      test.deepEqual(BatchPositionModel.updateParticipantPosition.getCalls().map(call => call.args), [
+        [{}, 7, 0, 0],
+        [{}, 15, 0, 0]
       ], 'updateParticipantPosition should be called with the expected arguments')
 
       // TODO: Assert on DB bulk insert of transferStateChanges in each function call
@@ -466,6 +622,14 @@ Test('BinProcessor', async (binProcessorTest) => {
 
       test.equal(spyCb.callCount, 10, 'callback should be called 10 times')
       test.equal(errorCb.callCount, 2, 'error callback should be called 2 times')
+      test.end()
+    })
+    iterateThroughBinsTest.test('iterateThroughBins should not affect if callback function throws error', async (test) => {
+      const spyCb = sandbox.stub()
+      spyCb.onFirstCall().throws()
+      await BinProcessor.iterateThroughBins(sampleBins, spyCb)
+
+      test.equal(spyCb.callCount, 10, 'callback should be called 10 times')
       test.end()
     })
     iterateThroughBinsTest.end()
